requirements.txt
# Core deep learning
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.36.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Data / scientific
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
